# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CommonResponse-any:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  model.CommonResponse-model_SettingResponse:
    properties:
      data:
        $ref: '#/definitions/model.SettingResponse'
      error:
        type: string
      success:
        type: boolean
    type: object
...
  model.FrontendTemplate:
    properties:
      author:
        type: string
      is_admin:
        type: boolean
      is_official:
        type: boolean
      name:
        type: string
      path:
        type: string
      repository:
        type: string
      version:
        type: string
    type: object
...
  model.Setting:
    properties:
      admin_template:
        type: string
      agent_real_ip_header:
        description: Agent真实IP
        type: string
      cover:
        description: 覆盖范围（0:提醒未被 IgnoredIPNotification 包含的所有服务器; 1:仅提醒被 IgnoredIPNotification
          包含的服务器;）
        type: integer
      custom_code:
        type: string
      custom_code_dashboard:
        type: string
      dns_servers:
        type: string
      enable_ip_change_notification:
        description: IP变更提醒
        type: boolean
      enable_plain_ip_in_notification:
        description: 通知信息IP不打码
        type: boolean
      ignored_ip_notification:
        description: 特定服务器IP（多个服务器用逗号分隔）
        type: string
      ignored_ip_notification_server_ids:
        additionalProperties:
          type: boolean
        type: object
      install_host:
        type: string
      ip_change_notification_group_id:
        type: integer
      language:
        description: 系统语言，默认 zh_CN
        type: string
      oauth2_providers:
        items:
          type: string
        type: array
      site_name:
        type: string
      tls:
        description: 用于前端判断生成的安装命令是否启用 TLS
        type: boolean
      user_template:
        type: string
      web_real_ip_header:
        description: 前端真实IP
        type: string
    type: object
  model.SettingForm:
    properties:
      agent_real_ip_header:
        description: Agent真实IP
        type: string
      cover:
        type: integer
      custom_code:
        type: string
      custom_code_dashboard:
        type: string
      dns_servers:
        type: string
      enable_ip_change_notification:
        type: boolean
      enable_plain_ip_in_notification:
        type: boolean
      ignored_ip_notification:
        type: string
      install_host:
        type: string
      ip_change_notification_group_id:
        description: IP变更提醒的通知组
        type: integer
      language:
        minLength: 2
        type: string
      site_name:
        minLength: 1
        type: string
      tls:
        type: boolean
      user_template:
        type: string
      web_real_ip_header:
        description: 前端真实IP
        type: string
    type: object
  model.SettingResponse:
    properties:
      config:
        $ref: '#/definitions/model.Setting'
      frontend_templates:
        items:
          $ref: '#/definitions/model.FrontendTemplate'
        type: array
      version:
        type: string
    type: object
...
host: localhost:8008
info:
  contact:
    email: hi@nai.ba
    name: API Support
    url: http://nezhahq.github.io
  description: Nezha Monitoring API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://nezhahq.github.io
  title: Nezha Monitoring API
  version: "1.0"
paths:
...
  /setting:
    get:
      description: List settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-model_SettingResponse'
      security:
      - BearerAuth: []
      summary: List settings
      tags:
      - common
    patch:
      consumes:
      - application/json
      description: Edit config
      parameters:
      - description: SettingForm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SettingForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Edit config
      tags:
      - admin required
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"